requirements.txt
pillow>=12.0.0
torch==2.5.1
numpy>=2.3.5
cog
torchvision
transformers @ git+https://github.com/huggingface/transformers.git